cli.go: Change config path
Check for legacy configs in UserHomeDir before using the UserConfigDir (go v1.13).
The UserConfigDir respects the XDG_CONFIG_HOME environment veriable as well as
Windows %APPDATA% and MacOS Library hierarchy.

This function takes advantage of APPDATA path for Windows, Library directory
for MacOS, and XDG_CONFIG_HOME for any platforms that export the varaible in
the shell.